+++ rust/noosphere-cli/src/native/mod.rs
@@ -11,11 +11,7 @@ pub mod workspace;
 #[cfg(any(test, feature = "helpers"))]
 pub mod helpers;
 
-use anyhow::Result;
-
-use noosphere_core::tracing::initialize_tracing;
-
-use clap::Parser;
+use std::path::{Path, PathBuf};
 
 use self::{
     cli::{AuthCommand, Cli, ConfigCommand, FollowCommand, KeyCommand, OrbCommand, SphereCommand},
@@ -29,24 +25,61 @@ use self::{
     },
     workspace::Workspace,
 };
+use anyhow::Result;
+use clap::Parser;
+use noosphere_core::tracing::initialize_tracing;
+use noosphere_storage::StorageConfig;
+
+/// Additional context used to invoke a [Cli] command.
+pub struct CliContext<'a> {
+    /// Path to the current working directory.
+    cwd: PathBuf,
+    /// Path to the global configuration directory, if provided.
+    global_config_directory: Option<&'a Path>,
+}
 
 #[cfg(not(doc))]
 #[allow(missing_docs)]
 pub async fn main() -> Result<()> {
     initialize_tracing(None);
-
-    let workspace = Workspace::new(&std::env::current_dir()?, None)?;
-
-    invoke_cli(Cli::parse(), workspace).await
+    let context = CliContext {
+        cwd: std::env::current_dir()?,
+        global_config_directory: None,
+    };
+    invoke_cli(Cli::parse(), &context).await
 }
 
-/// Invoke the CLI implementation imperatively. This is the entrypoint used by
-/// orb when handling a command line invocation. The [Cli] is produced by
-/// parsing the command line arguments, and the [Workspace] is initialized from
-/// the current working directory. The effect of invoking the CLI this way will
-/// depend on the [Cli] and [Workspace] provided (results may vary significantly
-/// depending on those inputs).
-pub async fn invoke_cli(cli: Cli, mut workspace: Workspace) -> Result<()> {
+/// Invoke the CLI implementation imperatively.
+///
+/// This is the entrypoint used by orb when handling a command line invocation.
+/// The [Cli] is produced by parsing the command line arguments, and internally
+/// creates a new [Workspace] from the current working directory.
+///
+/// Use [invoke_cli_with_workspace] if using your own [Workspace].
+pub async fn invoke_cli<'a>(cli: Cli, context: &CliContext<'a>) -> Result<()> {
+    let storage_config = if let OrbCommand::Serve {
+        storage_memory_cache_limit,
+        ..
+    } = &cli.command
+    {
+        Some(StorageConfig {
+            memory_cache_limit: *storage_memory_cache_limit,
+        })
+    } else {
+        None
+    };
+    let workspace = Workspace::new(
+        &context.cwd,
+        context.global_config_directory,
+        storage_config,
+    )?;
+
+    invoke_cli_with_workspace(cli, workspace).await
+}
+
+/// Same as [invoke_cli], but enables the caller to provide their own
+/// initialized [Workspace]
+pub async fn invoke_cli_with_workspace(cli: Cli, mut workspace: Workspace) -> Result<()> {
     match cli.command {
         OrbCommand::Key { command } => match command {
             KeyCommand::Create { name } => key_create(&name, &workspace).await?,
@@ -117,6 +150,7 @@ pub async fn invoke_cli(cli: Cli, mut workspace: Workspace) -> Result<()> {
             name_resolver_api,
             interface,
             port,
+            ..
         } => {
             serve(
                 interface,
@@ -124,11 +158,11 @@ pub async fn invoke_cli(cli: Cli, mut workspace: Workspace) -> Result<()> {
                 ipfs_api,
                 name_resolver_api,
                 cors_origin,
-                &workspace,
+                &mut workspace,
             )
-            .await?
+            .await?;
         }
-    };
+    }
 
     Ok(())
 }